add embed utils
