+++ cogs/fun/cog.py
@@ -252,7 +252,7 @@ class Fun(Base, commands.Cog):
         )
         embed.set_image(url=f"attachment://{str(index)}.png")
 
-        utils.general.add_author_footer(embed, inter.author, additional_text=[f" (hláškaid: #{str(index)})"])
+        utils.embed.add_author_footer(embed, inter.author, additional_text=[f" (hláškaid: #{str(index)})"])
 
         with open(fuchs_path + str(index) + ".png", "rb") as fp:
             await inter.send(embed=embed, file=disnake.File(fp=fp, filename=str(index) + ".png"))
