+++ utils/__init__.py
@@ -1,5 +1,6 @@
 from . import (
     constants,  # noqa: F401
+    embed,  # noqa: F401
     general,  # noqa: F401
     user,  # noqa: F401
 )
